Extract shared platform helpers into platform-utils.js and add Copilot plugin installation
Agent-Logs-Url: https://github.com/JohnLudlow/agents/sessions/acd3fd9a-8beb-43c6-bb33-71a0e330ca91
