+++ scripts/install.js
@@ -11,36 +11,21 @@
 
 const fs = require("fs");
 const path = require("path");
-const os = require("os");
+const { spawnSync } = require("child_process");
+
+const {
+  BACKUP_SUFFIX,
+  PLATFORMS,
+  getInstallMode,
+  getTargetDirectory,
+  removeDirectory,
+  copyDirectory,
+  backupExistingDirectory,
+} = require("./platform-utils.js");
 
 // ============================================================================
 // CONFIGURATION
 // ============================================================================
-
-const BACKUP_SUFFIX = `.johnludlow-backup-${new Date().toISOString().replace(/[:.]/g, "-")}`;
-
-// Installation platform definitions
-const PLATFORMS = {
-  opencode: {
-    name: "OpenCode",
-    emoji: "⚙️ ",
-    globalDir: path.join(os.homedir(), ".config", "opencode"),
-    localDir: (cwd) => path.join(cwd, ".opencode"),
-    requiresConfig: true,
-    configFile: "config.json",
-  },
-  copilot: {
-    name: "GitHub Copilot",
-    emoji: "🔌",
-    globalDir: path.join(os.homedir(), ".copilot"),
-    localDir: (cwd) => path.join(cwd, ".github"),
-    requiresConfig: false,
-    // For local mode, only manage these subdirectories
-    // to avoid affecting workflows, issue templates, and other .github metadata
-    // Only backup/restore these subdirs in local mode to avoid clobbering other .github content
-    localManagedSubDirs: ["agents", "skills"],
-  },
-};
 
 // Source directories (relative to package root)
 const SOURCE_DIRS = {
@@ -82,74 +67,6 @@ function getSourceSkillsDir(platform) {
 }
 
 /**
- * Copy directory recursively, skipping README.md files
- */
-function copyDirectory(source, target) {
-  if (!fs.existsSync(source)) {
-    return;
-  }
-
-  if (!fs.existsSync(target)) {
-    fs.mkdirSync(target, { recursive: true });
-  }
-
-  const files = fs.readdirSync(source);
-
-  files.forEach((file) => {
-    // Never copy README.md into agent/skill directories
-    if (file === "README.md") return;
-
-    const sourcePath = path.join(source, file);
-    const targetPath = path.join(target, file);
-
-    if (fs.statSync(sourcePath).isDirectory()) {
-      copyDirectory(sourcePath, targetPath);
-    } else {
-      fs.copyFileSync(sourcePath, targetPath);
-    }
-  });
-}
-
-/**
- * Backup existing directory or subdirectories for local Copilot mode
- *
- * For Copilot (both local and global): backs up only managed subdirectories individually
- * to avoid affecting workflows, issue templates, and other platform metadata
- * 
- * For other platforms: backs up by copying the entire directory
- * Uses copy instead of move to preserve originals during backup
- */
-function backupExistingDirectory(targetDir, platform, mode) {
-  const platformConfig = PLATFORMS[platform];
-  const isCopilot = platform === "copilot";
-  const managedSubDirs = isCopilot ? platformConfig.localManagedSubDirs : null;
-
-  if (isCopilot && managedSubDirs) {
-    // For Copilot (both local and global): backup only managed subdirectories individually
-    // This preserves other content like config.json, IDE settings, plugins, etc.
-    const backups = [];
-    for (const subDir of managedSubDirs) {
-      const subDirPath = path.join(targetDir, subDir);
-      if (fs.existsSync(subDirPath)) {
-        const backupDir = subDirPath + BACKUP_SUFFIX;
-        // Copy directory recursively instead of moving
-        copyDirectory(subDirPath, backupDir);
-        backups.push(backupDir);
-      }
-    }
-    return backups.length > 0 ? backups : null;
-  }
-
-  // For other platforms (OpenCode): backup entire directory by copying
-  if (fs.existsSync(targetDir)) {
-    const backupDir = targetDir + BACKUP_SUFFIX;
-    copyDirectory(targetDir, backupDir);
-    return [backupDir];
-  }
-  return null;
-}
-
-/**
  * Count markdown files in directory
  */
 function countMarkdownFiles(dir) {
@@ -163,53 +80,6 @@ function countMarkdownFiles(dir) {
 function checkOpenCodeInstalled() {
   const openCodeDir = PLATFORMS.opencode.globalDir;
   return fs.existsSync(openCodeDir);
-}
-
-/**
- * Determine installation mode (global vs local)
- */
-function getInstallMode() {
-  const isGlobalInstall = process.env.npm_config_global === "true";
-  return isGlobalInstall ? "global" : "local";
-}
-
-/**
- * Get target directory for a platform and mode
- * Note: Does NOT create the directory - that's done after backup
- */
-function getTargetDirectory(platform, mode, cwd = process.cwd()) {
-  const platformConfig = PLATFORMS[platform];
-  if (!platformConfig) {
-    throw new Error(`Unknown platform: ${platform}`);
-  }
-
-  if (mode === "global") {
-    return platformConfig.globalDir;
-  } else {
-    return platformConfig.localDir(cwd);
-  }
-}
-
-/**
- * Remove directory recursively
- */
-function removeDirectory(dir) {
-  if (!fs.existsSync(dir)) {
-    return;
-  }
-
-  const files = fs.readdirSync(dir);
-
-  for (const file of files) {
-    const filePath = path.join(dir, file);
-    if (fs.statSync(filePath).isDirectory()) {
-      removeDirectory(filePath);
-    } else {
-      fs.unlinkSync(filePath);
-    }
-  }
-
-  fs.rmdirSync(dir);
 }
 
 // ============================================================================
@@ -314,6 +184,101 @@ function installOpenCodeConfig(mode) {
   return targetDir;
 }
 
+// ============================================================================
+// PLUGIN INSTALLATION
+// ============================================================================
+
+// Recommended GitHub Copilot plugins (name@scope format)
+const COPILOT_PLUGINS = [
+  { name: "awesome-copilot", scope: "awesome-copilot", version: "v1.1.0" },
+  { name: "azure", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "doublecheck", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "dotnet", scope: "awesome-copilot", version: "v0.1.0" },
+  { name: "dotnet-diag", scope: "awesome-copilot", version: "v0.1.0" },
+  { name: "context-engineering", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "csharp-dotnet-development", scope: "awesome-copilot", version: "v1.1.0" },
+  { name: "csharp-mcp-development", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "devops-oncall", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "technical-spike", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "microsoft-docs", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "openapi-to-application-csharp-dotnet", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "polyglot-test-agent", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "roundup", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "project-planning", scope: "awesome-copilot", version: "v1.0.0" },
+  { name: "security-best-practices", scope: "awesome-copilot", version: "v1.0.0" },
+];
+
+/**
+ * Check whether the GitHub CLI (`gh`) is available on PATH
+ */
+function isGhCliAvailable() {
+  const result = spawnSync("gh", ["--version"], { encoding: "utf8", stdio: "pipe" });
+  return result.status === 0;
+}
+
+/**
+ * Attempt to install a single Copilot plugin via `gh extension install`
+ * Returns true on success, false on failure.
+ */
+function installPluginViaGh(scope, name) {
+  const result = spawnSync(
+    "gh",
+    ["extension", "install", `${scope}/${name}`],
+    { encoding: "utf8", stdio: "pipe" }
+  );
+  return result.status === 0;
+}
+
+/**
+ * Install recommended Copilot plugins.
+ *
+ * If the `gh` CLI is available the script attempts to install each plugin via
+ * `gh extension install <scope>/<name>`.  When `gh` is not available (or a
+ * plugin install fails) the plugin is flagged for manual installation and a
+ * human-readable summary is printed at the end.
+ */
+function installCopilotPlugins() {
+  console.log("\n🔌 Installing recommended Copilot plugins...");
+
+  const ghAvailable = isGhCliAvailable();
+  if (!ghAvailable) {
+    console.log("   ℹ️  GitHub CLI (gh) not found — plugins must be installed manually.");
+    console.log("      Install the gh CLI: https://cli.github.com/\n");
+    console.log("   Recommended plugins:");
+    COPILOT_PLUGINS.forEach(({ name, scope, version }) => {
+      console.log(`     • ${name}@${scope} (${version})`);
+      console.log(`       gh extension install ${scope}/${name}`);
+    });
+    return;
+  }
+
+  const succeeded = [];
+  const failed = [];
+
+  for (const { name, scope, version } of COPILOT_PLUGINS) {
+    process.stdout.write(`   → ${name}@${scope} (${version})... `);
+    if (installPluginViaGh(scope, name)) {
+      process.stdout.write("✓\n");
+      succeeded.push({ name, scope });
+    } else {
+      process.stdout.write("✗\n");
+      failed.push({ name, scope, version });
+    }
+  }
+
+  if (succeeded.length > 0) {
+    console.log(`\n   ✓ Installed ${succeeded.length} plugin(s) via gh extension`);
+  }
+
+  if (failed.length > 0) {
+    console.log(`\n   ⚠️  ${failed.length} plugin(s) could not be installed automatically:`);
+    failed.forEach(({ name, scope, version }) => {
+      console.log(`     • ${name}@${scope} (${version})`);
+      console.log(`       gh extension install ${scope}/${name}`);
+    });
+  }
+}
+
 /**
  * Display next steps
  */
@@ -376,6 +341,9 @@ async function main() {
     installOpenCodeConfig(mode);
     console.log("");
     const copilotPath = installPlatform("copilot", mode);
+
+    // Install recommended Copilot plugins
+    installCopilotPlugins();
 
     // Display next steps
     displayNextSteps(opencodePath, copilotPath, mode);
