+++ scripts/uninstall.js
@@ -4,134 +4,24 @@
  * @johnludlow/agents Uninstallation Script
  *
  * Removes agents and skills from both OpenCode and GitHub Copilot configurations.
- * Uses unified platform logic shared with install.js and restore.js.
+ * Uses platform helpers shared with install.js and restore.js via platform-utils.js.
  */
 
 const fs = require("fs");
 const path = require("path");
-const os = require("os");
 
-// ============================================================================
-// CONFIGURATION
-// ============================================================================
-
-const BACKUP_SUFFIX = `.johnludlow-backup-${new Date().toISOString().replace(/[:.]/g, "-")}`;
-
-// Installation platform definitions
-const PLATFORMS = {
-  opencode: {
-    name: "OpenCode",
-    emoji: "⚙️ ",
-    globalDir: path.join(os.homedir(), ".config", "opencode"),
-    localDir: (cwd) => path.join(cwd, ".opencode"),
-  },
-  copilot: {
-    name: "GitHub Copilot",
-    emoji: "🔌",
-    globalDir: path.join(os.homedir(), ".copilot"),
-    localDir: (cwd) => path.join(cwd, ".github"),
-    // For local mode, only manage these subdirectories
-    // to avoid affecting workflows, issue templates, and other .github metadata
-    // Only uninstall these subdirs in local mode to avoid removing other .github content
-    localManagedSubDirs: ["agents", "skills"],
-  },
-};
+const {
+  PLATFORMS,
+  getInstallMode,
+  getTargetDirectory,
+  removeDirectory,
+  listBackupsForPlatform,
+  findLatestBackup,
+} = require("./platform-utils.js");
 
 // ============================================================================
 // UTILITIES
 // ============================================================================
-
-/**
- * Determine installation mode (global vs local)
- */
-function getInstallMode() {
-  const isGlobalInstall = process.env.npm_config_global === "true";
-  return isGlobalInstall ? "global" : "local";
-}
-
-/**
- * Get target directory for a platform and mode
- */
-function getTargetDirectory(platform, mode, cwd = process.cwd()) {
-  const platformConfig = PLATFORMS[platform];
-  if (!platformConfig) {
-    throw new Error(`Unknown platform: ${platform}`);
-  }
-
-  if (mode === "global") {
-    return platformConfig.globalDir;
-  } else {
-    return platformConfig.localDir(cwd);
-  }
-}
-
-/**
- * Remove directory recursively
- */
-function removeDirectory(dir) {
-  if (!fs.existsSync(dir)) {
-    return;
-  }
-
-  const files = fs.readdirSync(dir);
-
-  for (const file of files) {
-    const filePath = path.join(dir, file);
-    if (fs.statSync(filePath).isDirectory()) {
-      removeDirectory(filePath);
-    } else {
-      fs.unlinkSync(filePath);
-    }
-  }
-
-  fs.rmdirSync(dir);
-}
-
-/**
- * Find and restore latest backup for a platform
- */
-function findLatestBackup(targetDir) {
-  const parentDir = path.dirname(targetDir);
-  if (!fs.existsSync(parentDir)) {
-    return null;
-  }
-
-  const dirName = path.basename(targetDir);
-  const files = fs.readdirSync(parentDir);
-  const backups = files
-    .filter(f => f.startsWith(dirName + ".johnludlow-backup-"))
-    .sort()
-    .reverse();
-
-  if (backups.length === 0) {
-    return null;
-  }
-
-  return path.join(parentDir, backups[0]);
-}
-
-/**
- * List all available backups for a platform
- */
-function listBackupsForPlatform(targetDir) {
-  const parentDir = path.dirname(targetDir);
-  if (!fs.existsSync(parentDir)) {
-    return [];
-  }
-
-  const dirName = path.basename(targetDir);
-  const files = fs.readdirSync(parentDir);
-  const backups = files
-    .filter(f => f.startsWith(dirName + ".johnludlow-backup-"))
-    .sort()
-    .reverse();
-
-  return backups.map(b => ({
-    name: b,
-    path: path.join(parentDir, b),
-    timestamp: b.replace(dirName + ".johnludlow-backup-", "")
-  }));
-}
 
 /**
  * Remove empty parent directory if possible
